cleanup resolver API
