+++ src/main/resolver/index.ts
@@ -1,18 +1,23 @@
+import { exportsForFile } from './exportsForFile'
+import { identifiersForStatements } from './identifiersForStatements'
+import { includesForFile } from './includesForFile'
+import { namespaceForFile } from './namespaceForFile'
+import { namespaceForInclude } from './namespaceForInclude'
+import { organizeByNamespace } from './organizeByNamespace'
+import { resolveConstValue } from './resolveConstValue'
+import { resolveIdentifierDefinition } from './resolveIdentifierDefinition'
+import { resolveIdentifierName } from './resolveIdentifierName'
 import { resolveNamespace } from './resolveNamespace'
 
-import {
-    exportsForFile,
-    identifiersForStatements,
-    resolveConstValue,
-    resolveIdentifierDefinition,
-    resolveIdentifierName,
-} from './utils'
-
 export const Resolver = {
-    resolveNamespace,
     exportsForFile,
-    resolveIdentifierDefinition,
+    identifiersForStatements,
+    includesForFile,
+    namespaceForFile,
+    namespaceForInclude,
+    organizeByNamespace,
     resolveConstValue,
     resolveIdentifierName,
-    identifiersForStatements,
+    resolveIdentifierDefinition,
+    resolveNamespace,
 }
